requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
greenlet
httpx
requests
pytest-asyncio